Took plot_format out of qha_yphon.py and put it into plotly_format.py
